requirements.txt
# AI Image Classification System - Python Dependencies
# Core ML Libraries
tensorflow>=2.8.0
numpy>=1.21.0
Pillow>=9.0.0

# Web Framework
Flask>=2.2.0
Werkzeug>=2.2.0

# Image Processing
opencv-python>=4.6.0

# Utilities
python-dotenv>=0.19.0

# Development (optional)
# flask-cors>=3.0.10
# gunicorn>=20.1.0
